Refactor stream client
