+++ pkg/agent/stream/streamer_test.go
